backend/requirements.txt
fastapi~=0.115.6
uvicorn
selenium~=4.27.1
python-dotenv~=1.0.1
mysql-connector-python~=9.1.0
starlette~=0.41.3
pydantic~=2.10.5